wined3d: Control WINED3D_BUFFER_DISCARD from the worker thread
This is ugly. It would be much nicer if BUFFER_DISCARD were a main
thread side flag, set on map and unset on draws. buffer_preload should
not discard the GL BO. Instead buffer_swap_mem should take care of this.

When implementing this, take care to keep the index buffer minding
intact. wined3d_buffer_preload should not touch BUFFER_DISCARD at all.
Instead, draws should remove it just like in this patch, just without
the ugly declaration of the flag in cs.c.

FIXME 2: The access to buffer->flags is completely unsynchronized. This
is asking for trouble too.